i965/gen9: Remove the halign/valign field setup code in fast copy blit
Experimentation with different values for src/dst horizontal/vertical
alignment showed that these fileds are not used on gen9 hardware.
So, this patch removes the code to setup those fields. A bug has been
created to get the spec language corrected.

Note: Setting up these fields isn't hurting on SKL. So, I'm also
fine with leaving them as it is until we see the corrections in
h/w spec.

Cc: Ben Widawsky <ben@bwidawsk.net>
Cc: Chad Versace <chad.versace@intel.com>
Signed-off-by: Anuj Phogat <anuj.phogat@gmail.com>