i965: Put the gt and slice count in the features macro
This saves a bit of typing for fields which we are obvious and always required
to be entered by the structure which is defining a platform. This is unlike
fields like URB sizes where the defaults might be fine.

Doing this also makes it easy and obvious to keep around preliminary hardware
information like we have for BXT.

This patch is just a preference of mine, so if people are opposed it doesn't
need to get merged.

Signed-off-by: Ben Widawsky <benjamin.widawsky@intel.com>
Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>