i965/cnl: Add a preliminary device for CNL
Since we've implemented all the known quirks for supporting gen10 with none of
the new features (ie. functions like Skylake), it should be safe to actually
enable the device.

As usual, thread counts and URB sizes are defines with conservative defaults
until permission is granted to publish the actual values. Note that WM thread
counts are strictly for scratch size calculation.

The PCI IDs in this patch were taken from the kernel patch which was implemented
first. That, along with the libdrm patch, will need to be verified before
posting this patch.

v2: rebased on some upstream changes

Signed-off-by: Ben Widawsky <benjamin.widawsky@intel.com>