i965/cnl/PO: implement temporary CNL A0 RO invalidate w/a
We could use the revision ID check and turn this into a proper patch, but this
shouldn't get upstreamed as A0 will not see the light of day.

Project GEN10:BUG:1940507 [CNL]
In A-step, Read-Only Invalidate does not flush HDC L1 Cache. Workaround is to
also set PIPE_CONTROL DC_FLUSH. Fixed in B-step.

Not-Signed-off-by: Ben Widawsky <benjamin.widawsky@intel.com>